# dockstore-webservice/src/main/resources/openapi3/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  contact:
    email: theglobalalliance@genomicsandhealth.org
    name: Dockstore@ga4gh
    url: https://discuss.dockstore.org/t/opening-helpdesk-tickets/1506
  description: The Dockstore API, includes proprietary and GA4GH V1+V2 endpoints
  license:
    name: Apache 2.0
    url: https://github.com/dockstore/dockstore/blob/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Dockstore
  version: 1.6.0
openapi: 3.0.1
paths:
...
  /api/ga4gh/v2/extended/tools/entry/_search:
    post:
      operationId: toolsIndexSearch
      requestBody:
        content:
          '*/*':
            schema:
              type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
...
  /metadata/runner_dependencies:
    get:
      description: Returns the file containing runner dependencies, NO authentication
      operationId: getRunnerDependencies
      parameters:
      - description: The Dockstore client version
        in: query
        name: client_version
        schema:
          type: string
      - description: Python version, only relevant for the cwltool runner
        in: query
        name: python_version
        schema:
          default: "2"
          type: string
      - description: The tool runner
        in: query
        name: runner
        schema:
          default: cwltool
          enum:
          - cwltool
          type: string
      - description: Response type
        in: query
        name: output
        schema:
          default: text
          enum:
          - json
          - text
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                type: string
          description: The requirements.txt file
      summary: Returns the file containing runner dependencies
      tags:
      - metadata
...